api/requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
black==21.5b1
    # via -r requirements-dev.in
boto3==1.18.46
    # via -r requirements.in
botocore==1.21.46
    # via
    #   boto3
    #   s3transfer
brotlipy==0.7.0
    # via -r requirements.in
cffi==1.14.6
    # via brotlipy
click==8.0.1
    # via
    #   black
    #   pip-tools
coverage==5.5
    # via pytest-cov
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django-ckeditor==6.1.0
    # via -r requirements.in
django-configurations==2.2
    # via -r requirements.in
django-cors-headers==3.8.0
    # via -r requirements.in
django-import-export==2.6.0
    # via -r requirements.in
django-jquery==3.1.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-ckeditor
django-storages==1.11.1
    # via -r requirements.in
django-taggit-serializer==0.1.7
    # via -r requirements.in
django-taggit==1.5.1
    # via
    #   -r requirements.in
    #   django-taggit-serializer
django==3.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-import-export
    #   django-jquery
    #   django-storages
    #   django-taggit
    #   djangorestframework
    #   drf-spectacular
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.19.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
faker==8.13.2
    # via -r requirements-dev.in
fancycompleter==0.9.1
    # via pdbpp
flake8==3.9.2
    # via pytest-flake8
gunicorn==20.1.0
    # via -r requirements.in
importlib-metadata==4.8.1
    # via
    #   click
    #   flake8
    #   jsonschema
    #   pep517
    #   pluggy
    #   pytest
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
ipython==7.27.0
    # via -r requirements-dev.in
jedi==0.18.0
    # via ipython
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via drf-spectacular
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pdbpp==0.10.3
    # via -r requirements-dev.in
pep517==0.11.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.2
    # via -r requirements.in
pip-tools==6.3.0
    # via -r requirements-dev.in
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via ipython
psycopg2==2.9.1
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via flake8
pygments==2.10.0
    # via
    #   ipython
    #   pdbpp
pyparsing==2.4.7
    # via packaging
pyrepl==0.9.0
    # via fancycompleter
pyrsistent==0.18.0
    # via jsonschema
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-django==4.4.0
    # via -r requirements-dev.in
pytest-flake8==1.0.7
    # via -r requirements-dev.in
pytest==6.2.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-flake8
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
pyyaml==5.4.1
    # via
    #   drf-spectacular
    #   tablib
regex==2021.8.28
    # via black
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via
    #   django-configurations
    #   django-taggit-serializer
    #   jsonschema
    #   python-dateutil
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.0.0
    # via
    #   -r requirements.in
    #   django-import-export
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pytest
    #   pytest-cov
tomli==1.2.1
    # via pep517
tqdm==4.62.3
    # via -r requirements-dev.in
traitlets==5.1.0
    # via
    #   ipython
    #   matplotlib-inline
typed-ast==1.4.3
    # via black
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   black
    #   importlib-metadata
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.7
    # via botocore
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
whitenoise==5.3.0
    # via -r requirements.in
wmctrl==0.4
    # via pdbpp
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zipp==3.5.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
